requirements.txt
langchain==0.3.25
langchain-openai==0.3.23
langchain-community==0.3.25
langchain-core==0.3.65
langchain-text-splitters==0.3.8
faiss-cpu==1.11.0
openai==1.88.0
python-dotenv==1.0.0
pandas==2.3.0
fastapi==0.115.12
uvicorn==0.34.3
python-multipart==0.0.20
streamlit==1.45.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
pydantic>=2.0.0,<3.0.0
pydantic-settings==2.1.0
tiktoken==0.5.1
packaging==23.2
langsmith==0.0.77
